+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: nbudzins <nbudzins@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 23:50:48 by nbudzins          #+#    #+#             */
-/*   Updated: 2024/02/28 20:59:13 by nbudzins         ###   ########.fr       */
+/*   Updated: 2024/02/29 17:21:52 by nbudzins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while(i < n)
